+++ services/inc/time_util.h
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2023 Particle Industries, Inc.  All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation, either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include <cstdint>
+
+#include "rtc_hal.h"
+
+#include "system_error.h"
+
+namespace particle {
+
+/**
+ * Get the number of milliseconds elapsed since the Unix epoch.
+ *
+ * @return Number of milliseconds on success, otherwise an error code defined by `system_error_t`.
+ */
+inline int64_t getMillisSinceEpoch() {
+    if (!hal_rtc_time_is_valid(nullptr)) {
+        return SYSTEM_ERROR_HAL_RTC_INVALID_TIME;
+    }
+    timeval tv = {};
+    int r = hal_rtc_get_time(&tv, nullptr);
+    if (r < 0) {
+        return r;
+    }
+    return tv.tv_sec * 1000ll + tv.tv_usec / 1000;
+}
+
+} // namespace particle
