+++ system/src/ledger/ledger_manager.cpp
@@ -25,99 +25,45 @@
 
 #include <algorithm>
 #include <mutex>
-#include <cstdio>
 #include <cstring>
-#include <cstdarg>
 #include <cassert>
 
 #include <pb_decode.h>
 
-#include "system_ledger_internal.h"
-
 #include "control/common.h" // FIXME: Move Protobuf utilities to another directory
+#include "ledger.h"
+#include "ledger_manager.h"
+#include "ledger_util.h"
+#include "system_ledger.h"
 
 #include "timer_hal.h"
-#include "rtc_hal.h"
 
 #include "nanopb_misc.h"
 #include "file_util.h"
+#include "time_util.h"
 #include "endian_util.h"
+#include "c_string.h"
 #include "scope_guard.h"
 #include "check.h"
 
-#include "cloud/cloud.pb.h" // Cloud protocol definitions
-#include "cloud/ledger.pb.h" // Cloud protocol definitions
-#include "ledger.pb.h" // Internal definitions
+#include "cloud/cloud.pb.h"
+#include "cloud/ledger.pb.h"
 
 #define PB_CLOUD(_name) particle_cloud_##_name
 #define PB_LEDGER(_name) particle_cloud_ledger_##_name
-#define PB_INTERNAL(_name) particle_firmware_##_name
 
 LOG_SOURCE_CATEGORY("system.ledger");
-
-static_assert(LEDGER_SCOPE_UNKNOWN == (int)PB_LEDGER(Scope_SCOPE_UNKNOWN) &&
-        LEDGER_SCOPE_DEVICE == (int)PB_LEDGER(Scope_SCOPE_DEVICE) &&
-        LEDGER_SCOPE_PRODUCT == (int)PB_LEDGER(Scope_SCOPE_PRODUCT) &&
-        LEDGER_SCOPE_OWNER == (int)PB_LEDGER(Scope_SCOPE_OWNER));
-
-static_assert(LEDGER_SYNC_DIRECTION_UNKNOWN == (int)PB_LEDGER(SyncDirection_SYNC_DIRECTION_UNKNOWN) &&
-        LEDGER_SYNC_DIRECTION_DEVICE_TO_CLOUD == (int)PB_LEDGER(SyncDirection_SYNC_DIRECTION_DEVICE_TO_CLOUD) &&
-        LEDGER_SYNC_DIRECTION_CLOUD_TO_DEVICE == (int)PB_LEDGER(SyncDirection_SYNC_DIRECTION_CLOUD_TO_DEVICE));
 
 namespace particle {
 
 using control::common::EncodedString;
-using control::common::DecodedCString;
 using fs::FsLock;
 
 namespace system {
 
 namespace {
 
-/*
-    The ledger directory is organized as follows:
-
-    /usr/ledger/
-    |
-    +--- my-ledger/ - Files of the ledger called "my-ledger"
-    |    |
-    |    +--- staged/ - Fully written ledger data files
-    |    |    +--- 0001
-    |    |    +--- 0002
-    |    |    +--- ...
-    |    |
-    |    +--- temp/ - Partially written ledger data files
-    |    |    +--- 0003
-    |    |    +--- 0004
-    |    |    +--- ...
-    |    |
-    |    +--- current - Current ledger data
-    |
-    +--- ...
-
-    When a ledger is opened for writing, a temporary file for the new ledger data is created in the
-    "temp" directory. When writing is finished, if there are active readers of the ledger data, the
-    temporary file is moved to the "staged" directory, otherwise the file is moved to "current" to
-    replace the current ledger data.
-
-    When a ledger is opened for reading, it is checked whether the most recent ledger data is stored
-    in "current" or a file in the "staged" directory. The ledger data is then read from the respective
-    file.
-
-    The ledger instance tracks the readers of the current and staged data to ensure that the relevant
-    files are modified or removed only when the last reader accessing them is closed. When all readers
-    are closed, the most recent staged data is moved to "current" and all other files in "staged" are
-    removed.
-*/
-const auto LEDGER_ROOT_DIR = "/usr/ledger";
-const auto TEMP_DATA_DIR_NAME = "temp";
-const auto STAGED_DATA_DIR_NAME = "staged";
-const auto CURRENT_DATA_FILE_NAME = "current";
-
-const unsigned DATA_FORMAT_VERSION = 1;
-
-const size_t MAX_LEDGER_NAME_LEN = 32; // Must match the maximum length in ledger.proto
-const size_t MAX_PATH_LEN = 127;
+const unsigned MAX_LEDGER_COUNT = 20;
 
 const unsigned MIN_SYNC_DELAY = 5000;
 const unsigned MAX_SYNC_DELAY = 30000;
@@ -125,158 +71,7 @@ const unsigned MAX_SYNC_DELAY = 30000;
 const auto REQUEST_URI = "L";
 const auto REQUEST_METHOD = COAP_METHOD_POST;
 
-// Internal result codes
-enum Result {
-    RESULT_CURRENT_DATA_NOT_FOUND = 1,
-    RESULT_RESPONSE_SENT = 2
-};
-
-/*
-    The layout of a ledger data file:
-
-    Field     | Size      | Description
-    ----------+-----------+------------
-    data      | data_size | Contents of the ledger in an application-specific format
-    info      | info_size | Protobuf-encoded ledger info (particle.firmware.LedgerInfo)
-    data_size | 4         | Size of the "data" field (unsigned integer)
-    info_size | 4         | Size of the "info" field (unsigned integer)
-    version   | 4         | Format version number (unsigned integer)
-
-    All integer fields are encoded in little-endian byte order.
-*/
-struct LedgerDataFooter {
-    uint32_t dataSize;
-    uint32_t infoSize;
-    uint32_t version;
-} __attribute__((packed));
-
-int readFooter(lfs_t* fs, lfs_file_t* file, size_t* dataSize = nullptr, size_t* infoSize = nullptr, int* version = nullptr) {
-    LedgerDataFooter f = {};
-    size_t n = CHECK_FS(lfs_file_read(fs, file, &f, sizeof(f)));
-    if (n != sizeof(f)) {
-        LOG(ERROR, "Unexpected end of ledger data file");
-        return SYSTEM_ERROR_LEDGER_INVALID_FORMAT;
-    }
-    if (dataSize) {
-        *dataSize = littleEndianToNative(f.dataSize);
-    }
-    if (infoSize) {
-        *infoSize = littleEndianToNative(f.infoSize);
-    }
-    if (version) {
-        *version = littleEndianToNative(f.version);
-    }
-    return n;
-}
-
-int writeFooter(lfs_t* fs, lfs_file_t* file, size_t dataSize, size_t infoSize, int version = DATA_FORMAT_VERSION) {
-    LedgerDataFooter f = {};
-    f.dataSize = nativeToLittleEndian(dataSize);
-    f.infoSize = nativeToLittleEndian(infoSize);
-    f.version = nativeToLittleEndian(version);
-    size_t n = CHECK_FS(lfs_file_write(fs, file, &f, sizeof(f)));
-    if (n != sizeof(f)) {
-        LOG(ERROR, "Unexpected number of bytes written");
-        return SYSTEM_ERROR_FILESYSTEM;
-    }
-    return n;
-}
-
-int writeLedgerInfo(lfs_t* fs, lfs_file_t* file, const char* ledgerName, const LedgerInfo& info) {
-    PB_INTERNAL(LedgerInfo) pbInfo = {};
-    size_t n = strlcpy(pbInfo.name, ledgerName, sizeof(pbInfo.name));
-    if (n >= sizeof(pbInfo.name)) {
-        return SYSTEM_ERROR_INTERNAL; // The name is longer than specified in ledger.proto
-    }
-    pbInfo.scope = static_cast<PB_LEDGER(Scope)>(info.scope());
-    pbInfo.sync_direction = static_cast<PB_LEDGER(SyncDirection)>(info.syncDirection());
-    pbInfo.last_updated = info.lastUpdated();
-    pbInfo.last_synced = info.lastSynced();
-    pbInfo.sync_pending = info.syncPending();
-    n = CHECK(encodeProtobufToFile(file, &PB_INTERNAL(LedgerInfo_msg), &pbInfo));
-    return n;
-}
-
-int formatLedgerPath(char* buf, size_t size, const char* ledgerName, const char* fmt = nullptr, ...) {
-    // Format the prefix part of the path
-    int n = snprintf(buf, size, "%s/%s/", LEDGER_ROOT_DIR, ledgerName);
-    if (n < 0) {
-        return SYSTEM_ERROR_INTERNAL;
-    }
-    size_t pos = n;
-    if (pos >= size) {
-        return SYSTEM_ERROR_PATH_TOO_LONG;
-    }
-    if (fmt) {
-        // Format the rest of the path
-        va_list args;
-        va_start(args, fmt);
-        n = vsnprintf(buf + pos, size - pos, fmt, args);
-        va_end(args);
-        if (n < 0) {
-            return SYSTEM_ERROR_INTERNAL;
-        }
-        pos += n;
-        if (pos >= size) {
-            return SYSTEM_ERROR_PATH_TOO_LONG;
-        }
-    }
-    return pos;
-}
-
-inline int getLedgerDirPath(char* buf, size_t size, const char* ledgerName) {
-    CHECK(formatLedgerPath(buf, size, ledgerName));
-    return 0;
-}
-
-inline int getTempDirPath(char* buf, size_t size, const char* ledgerName) {
-    CHECK(formatLedgerPath(buf, size, ledgerName, "%s/", TEMP_DATA_DIR_NAME));
-    return 0;
-}
-
-inline int getTempFilePath(char* buf, size_t size, const char* ledgerName, int seqNum) {
-    CHECK(formatLedgerPath(buf, size, ledgerName, "%s/%04d", TEMP_DATA_DIR_NAME, seqNum));
-    return 0;
-}
-
-inline int getTempFilePath(char* buf, size_t size, const char* ledgerName, const char* fileName) {
-    CHECK(formatLedgerPath(buf, size, ledgerName, "%s/%s", TEMP_DATA_DIR_NAME, fileName));
-    return 0;
-}
-
-inline int getStagedDirPath(char* buf, size_t size, const char* ledgerName) {
-    CHECK(formatLedgerPath(buf, size, ledgerName, "%s/", STAGED_DATA_DIR_NAME));
-    return 0;
-}
-
-inline int getStagedFilePath(char* buf, size_t size, const char* ledgerName, int seqNum) {
-    CHECK(formatLedgerPath(buf, size, ledgerName, "%s/%04d", STAGED_DATA_DIR_NAME, seqNum));
-    return 0;
-}
-
-inline int getStagedFilePath(char* buf, size_t size, const char* ledgerName, const char* fileName) {
-    CHECK(formatLedgerPath(buf, size, ledgerName, "%s/%s", STAGED_DATA_DIR_NAME, fileName));
-    return 0;
-}
-
-inline int getCurrentFilePath(char* buf, size_t size, const char* ledgerName) {
-    CHECK(formatLedgerPath(buf, size, ledgerName, "%s", CURRENT_DATA_FILE_NAME));
-    return 0;
-}
-
-bool isLedgerNameValid(const char* name) {
-    size_t len = 0;
-    char c = 0;
-    while ((c = *name++)) {
-        if (!(++len <= MAX_LEDGER_NAME_LEN && ((c >= 'a' && c <= 'z') || (c >= '0' && c <= '9') || c == '-'))) {
-            return false;
-        }
-    }
-    if (len == 0) {
-        return false;
-    }
-    return true;
-}
+const size_t MAX_PATH_LEN = 127;
 
 int getStreamForSubmessage(coap_message* msg, pb_istream_t* stream, uint32_t fieldTag) {
     pb_istream_t s = {};
@@ -313,43 +108,20 @@ int getStreamForSubmessage(coap_message* msg, pb_istream_t* stream, uint32_t fie
     }
 }
 
-int64_t getCurrentTime() {
-    if (!hal_rtc_time_is_valid(nullptr)) {
-        return SYSTEM_ERROR_HAL_RTC_INVALID_TIME;
-    }
-    timeval tv = {};
-    CHECK(hal_rtc_get_time(&tv, nullptr));
-    return tv.tv_sec * 1000ll + tv.tv_usec / 1000;
-}
-
-// Helper functions that transform LittleFS errors to a system error
-inline int renameFile(lfs_t* fs, const char* oldPath, const char* newPath) {
-    CHECK_FS(lfs_rename(fs, oldPath, newPath));
-    return 0;
-}
-
-inline int removeFile(lfs_t* fs, const char* path) {
-    CHECK_FS(lfs_remove(fs, path));
-    return 0;
-}
-
-inline int closeFile(lfs_t* fs, lfs_file_t* file) {
-    CHECK_FS(lfs_file_close(fs, file));
-    return 0;
-}
-
-inline int closeDir(lfs_t* fs, lfs_dir_t* dir) {
+inline int closeDir(lfs_t* fs, lfs_dir_t* dir) { // Transforms the LittleFS error to a system error
     CHECK_FS(lfs_dir_close(fs, dir));
     return 0;
 }
 
 } // namespace
 
-struct LedgerManager::LedgerContext {
+namespace detail {
+
+struct LedgerSyncContext {
     CString name; // Ledger name
     Ledger* instance; // Ledger instance. If null, the ledger is not instantiated
     ledger_sync_direction syncDir; // Sync direction
-    int pendingState; // Pending state flags
+    int pendingState; // Pending state flags (see LedgerManager::PendingState)
     bool taskRunning; // Whether an asynchronous task is running for this ledger
     union {
         struct { // Fields specific to a device-to-cloud ledger
@@ -362,7 +134,7 @@ struct LedgerManager::LedgerContext {
         };
     };
 
-    LedgerContext() :
+    LedgerSyncContext() :
             instance(nullptr),
             syncDir(LEDGER_SYNC_DIRECTION_UNKNOWN),
             pendingState(0),
@@ -372,19 +144,21 @@ struct LedgerManager::LedgerContext {
             changed(false) {
     }
 
-    void resetDeviceToCloudSyncState() {
+    void resetDeviceToCloudState() {
         syncTime = 0;
         forcedSyncTime = 0;
         changed = false;
     }
 
-    void resetCloudToDeviceSyncState() {
+    void resetCloudToDeviceState() {
         lastUpdated = 0;
     }
 };
 
+} // namespace detail
+
 LedgerManager::LedgerManager() :
-        curLedger_(nullptr),
+        curCtx_(nullptr),
         nextSyncTime_(0),
         state_(State::NEW),
         pendingState_(0),
@@ -392,8 +166,10 @@ LedgerManager::LedgerManager() :
 }
 
 LedgerManager::~LedgerManager() {
-    coap_remove_request_handler(REQUEST_URI, REQUEST_METHOD, nullptr);
-    coap_remove_connection_handler(connectionCallback, nullptr);
+    if (state_ != State::NEW) {
+        coap_remove_request_handler(REQUEST_URI, REQUEST_METHOD, nullptr);
+        coap_remove_connection_handler(connectionCallback, nullptr);
+    }
 }
 
 int LedgerManager::init() {
@@ -401,7 +177,7 @@ int LedgerManager::init() {
         return SYSTEM_ERROR_INVALID_STATE;
     }
     // Enumerate local ledgers
-    LedgerContexts ledgers;
+    LedgerSyncContexts contexts;
     FsLock fs;
     lfs_dir_t dir = {};
     int r = lfs_dir_open(fs.instance(), &dir, LEDGER_ROOT_DIR);
@@ -421,6 +197,10 @@ int LedgerManager::init() {
             if (std::strcmp(entry.name, ".") == 0 || std::strcmp(entry.name, "..") == 0) {
                 continue;
             }
+            if (contexts.size() >= (int)MAX_LEDGER_COUNT) {
+                LOG(ERROR, "Maximum number of ledgers exceeded, skipping ledger: %s", entry.name);
+                continue;
+            }
             // Load the ledger info
             Ledger ledger;
             int r = ledger.init(entry.name);
@@ -429,7 +209,7 @@ int LedgerManager::init() {
                 continue;
             }
             // Create a sync context
-            std::unique_ptr<LedgerContext> ctx(new(std::nothrow) LedgerContext());
+            std::unique_ptr<LedgerSyncContext> ctx(new(std::nothrow) LedgerSyncContext());
             if (!ctx) {
                 return SYSTEM_ERROR_NO_MEMORY;
             }
@@ -444,7 +224,7 @@ int LedgerManager::init() {
             } else if (info.syncPending()) { // DEVICE_TO_CLOUD or UNKNOWN
                 ctx->changed = true;
             }
-            if (!ledgers.append(std::move(ctx))) {
+            if (!contexts.append(std::move(ctx))) {
                 return SYSTEM_ERROR_NO_MEMORY;
             }
         }
@@ -458,7 +238,7 @@ int LedgerManager::init() {
     });
     CHECK(coap_add_request_handler(REQUEST_URI, REQUEST_METHOD, requestCallback, this, nullptr));
     removeConnHandler.dismiss();
-    allLedgers_ = std::move(ledgers);
+    contexts_ = std::move(contexts);
     state_ = State::OFFLINE;
     return 0;
 }
@@ -470,19 +250,23 @@ int LedgerManager::getLedger(RefCountPtr<Ledger>& ledger, const char* name, bool
     }
     // Check if the requested ledger is already instantiated
     bool found = false;
-    auto it = findLedger(name, found);
+    auto it = findContext(name, found);
     if (found && (*it)->instance) {
         ledger = (*it)->instance;
         return 0;
     }
-    std::unique_ptr<LedgerContext> newCtx;
-    LedgerContext* ctx = it->get();
+    std::unique_ptr<LedgerSyncContext> newCtx;
+    LedgerSyncContext* ctx = nullptr;
     if (!found) {
         if (!create) {
             return SYSTEM_ERROR_LEDGER_NOT_FOUND;
         }
+        if (contexts_.size() >= (int)MAX_LEDGER_COUNT) {
+            LOG(ERROR, "Maximum number of ledgers exceeded");
+            return SYSTEM_ERROR_LEDGER_TOO_MANY;
+        }
         // Create a new sync context
-        newCtx.reset(new(std::nothrow) LedgerContext());
+        newCtx.reset(new(std::nothrow) LedgerSyncContext());
         if (!newCtx) {
             return SYSTEM_ERROR_NO_MEMORY;
         }
@@ -491,6 +275,8 @@ int LedgerManager::getLedger(RefCountPtr<Ledger>& ledger, const char* name, bool
             return SYSTEM_ERROR_NO_MEMORY;
         }
         ctx = newCtx.get();
+    } else {
+        ctx = it->get();
     }
     // Create a ledger instance
     auto lr = makeRefCountPtr<Ledger>(ctx);
@@ -503,8 +289,8 @@ int LedgerManager::getLedger(RefCountPtr<Ledger>& ledger, const char* name, bool
         return r;
     }
     if (!found) {
-        it = allLedgers_.insert(it, std::move(newCtx));
-        if (it == allLedgers_.end()) {
+        it = contexts_.insert(it, std::move(newCtx));
+        if (it == contexts_.end()) {
             return SYSTEM_ERROR_NO_MEMORY;
         }
         if (state_ >= State::READY) {
@@ -526,15 +312,15 @@ int LedgerManager::removeLedgerData(const char* name) {
         return SYSTEM_ERROR_INVALID_STATE;
     }
     bool found = false;
-    auto it = findLedger(name, found);
+    auto it = findContext(name, found);
     if (found) {
         if ((*it)->instance) {
             return SYSTEM_ERROR_LEDGER_IN_USE;
         }
-        allLedgers_.erase(it);
+        contexts_.erase(it);
     }
     char path[MAX_PATH_LEN + 1];
-    CHECK(getLedgerDirPath(path, sizeof(path), name));
+    CHECK(formatLedgerPath(path, sizeof(path), name));
     CHECK(rmrf(path));
     return 0;
 }
@@ -544,12 +330,12 @@ int LedgerManager::removeAllData() {
     if (state_ != State::OFFLINE) {
         return SYSTEM_ERROR_INVALID_STATE;
     }
-    for (auto& ctx: allLedgers_) {
+    for (auto& ctx: contexts_) {
         if (ctx->instance) {
             return SYSTEM_ERROR_LEDGER_IN_USE;
         }
     }
-    allLedgers_.clear();
+    contexts_.clear();
     CHECK(rmrf(LEDGER_ROOT_DIR));
     return 0;
 }
@@ -582,8 +368,8 @@ int LedgerManager::processTasks() {
         auto now = hal_timer_millis(nullptr);
         if (now >= nextSyncTime_) {
             uint64_t t = 0;
-            LedgerContext* ctx = nullptr;
-            for (auto& c: allLedgers_) {
+            LedgerSyncContext* ctx = nullptr;
+            for (auto& c: contexts_) {
                 if (c->pendingState & PendingState::SYNC_TO_CLOUD) {
                     if (now >= c->syncTime) {
                         ctx = c.get();
@@ -603,8 +389,8 @@ int LedgerManager::processTasks() {
         }
     }
     if (pendingState_ & PendingState::SYNC_FROM_CLOUD) {
-        LedgerContext* ctx = nullptr;
-        for (auto& c: allLedgers_) {
+        LedgerSyncContext* ctx = nullptr;
+        for (auto& c: contexts_) {
             if (c->pendingState & PendingState::SYNC_FROM_CLOUD) {
                 ctx = c.get();
                 break;
@@ -624,7 +410,7 @@ int LedgerManager::notifyConnected() {
         return SYSTEM_ERROR_INVALID_STATE;
     }
     LOG(TRACE, "Connected");
-    for (auto& ctx: allLedgers_) {
+    for (auto& ctx: contexts_) {
         switch (ctx->syncDir) {
         case LEDGER_SYNC_DIRECTION_DEVICE_TO_CLOUD: {
             if (ctx->changed) {
@@ -666,7 +452,7 @@ void LedgerManager::notifyDisconnected(int /* error */) {
         }
         stream_.reset();
     }
-    for (auto& ctx: allLedgers_) {
+    for (auto& ctx: contexts_) {
         if (ctx->syncDir == LEDGER_SYNC_DIRECTION_DEVICE_TO_CLOUD || ctx->syncDir == LEDGER_SYNC_DIRECTION_UNKNOWN) {
             ctx->syncTime = 0;
             ctx->forcedSyncTime = 0;
@@ -676,7 +462,7 @@ void LedgerManager::notifyDisconnected(int /* error */) {
     }
     pendingState_ = 0;
     nextSyncTime_ = 0;
-    curLedger_ = nullptr;
+    curCtx_ = nullptr;
     reqId_ = COAP_INVALID_REQUEST_ID;
     state_ = State::OFFLINE;
 }
@@ -734,7 +520,7 @@ int LedgerManager::receiveNotifyUpdateRequest(coap_message* msg, int /* reqId */
         }
         // Find the updated ledger
         bool found = false;
-        auto it = self->findLedger(pbLedger.name, found);
+        auto it = self->findContext(pbLedger.name, found);
         if (!found) {
             LOG(WARN, "Unknown ledger: %s", pbLedger.name);
             return true; // Ignore
@@ -764,7 +550,7 @@ int LedgerManager::receiveNotifyUpdateRequest(coap_message* msg, int /* reqId */
 
 int LedgerManager::receiveResetInfoRequest(coap_message* msg, int /* reqId */) {
     LOG(WARN, "Received a reset request, re-requesting ledger info");
-    for (auto& ctx: allLedgers_) {
+    for (auto& ctx: contexts_) {
         setPendingState(ctx.get(), PendingState::GET_INFO);
     }
     return 0;
@@ -836,9 +622,9 @@ int LedgerManager::receiveResponse(coap_message* msg, int code, int reqId) {
 }
 
 int LedgerManager::receiveSetDataResponse(coap_message* msg, int result, int /* reqId */) {
-    assert(state_ == State::SYNC_TO_CLOUD && curLedger_ && curLedger_->syncDir == LEDGER_SYNC_DIRECTION_DEVICE_TO_CLOUD &&
-            curLedger_->taskRunning);
-    auto name = (const char*)curLedger_->name;
+    assert(state_ == State::SYNC_TO_CLOUD && curCtx_ && curCtx_->syncDir == LEDGER_SYNC_DIRECTION_DEVICE_TO_CLOUD &&
+            curCtx_->taskRunning);
+    auto name = (const char*)curCtx_->name;
     if (result != 0) {
         LOG(ERROR, "Failed to sync ledger: %s; response result: %d", name, result);
         if (result != PB_CLOUD(Response_Result_LEDGER_NOT_FOUND) &&
@@ -847,38 +633,39 @@ int LedgerManager::receiveSetDataResponse(coap_message* msg, int result, int /* 
         }
         // Ledger is no longer accessible, re-request its info
         LOG(WARN, "Re-requesting ledger info: %s", name);
-        setPendingState(curLedger_, PendingState::GET_INFO | PendingState::SYNC_TO_CLOUD);
-        curLedger_->taskRunning = false;
-        curLedger_ = nullptr;
+        setPendingState(curCtx_, PendingState::GET_INFO | PendingState::SYNC_TO_CLOUD);
+        curCtx_->taskRunning = false;
+        curCtx_ = nullptr;
         return 0;
     }
     LOG(TRACE, "Sent ledger data: %s", name);
     LedgerInfo info;
-    auto t = CHECK(getCurrentTime());
-    info.lastSynced(t);
-    if (!(curLedger_->pendingState & PendingState::SYNC_TO_CLOUD)) {
+    auto now = CHECK(getMillisSinceEpoch());
+    info.lastSynced(now);
+    if (!(curCtx_->pendingState & PendingState::SYNC_TO_CLOUD)) {
         info.syncPending(false);
-        curLedger_->syncTime = 0;
-        curLedger_->forcedSyncTime = 0;
-        curLedger_->changed = false;
+        curCtx_->syncTime = 0;
+        curCtx_->forcedSyncTime = 0;
+        curCtx_->changed = false;
     } else {
         auto now = hal_timer_millis(nullptr);
-        curLedger_->syncTime = now + MIN_SYNC_DELAY;
-        curLedger_->forcedSyncTime = now + MAX_SYNC_DELAY;
+        curCtx_->syncTime = now + MIN_SYNC_DELAY;
+        curCtx_->forcedSyncTime = now + MAX_SYNC_DELAY;
     }
     RefCountPtr<Ledger> ledger;
-    CHECK(getLedger(ledger, curLedger_->name, false /* create */));
+    CHECK(getLedger(ledger, curCtx_->name));
     CHECK(ledger->updateInfo(info));
-    curLedger_->taskRunning = false;
-    curLedger_ = nullptr;
+    ledger->notifySynced(); // TODO: Invoke asynchronously
+    curCtx_->taskRunning = false;
+    curCtx_ = nullptr;
     // TODO: Reorder the ledger entries so that they're synchronized in a round-robin fashion
     return 0;
 }
 
 int LedgerManager::receiveGetDataResponse(coap_message* msg, int result, int /* reqId */) {
-    assert(state_ == State::SYNC_FROM_CLOUD && curLedger_ && curLedger_->syncDir == LEDGER_SYNC_DIRECTION_CLOUD_TO_DEVICE &&
-            curLedger_->taskRunning && !stream_);
-    auto name = (const char*)curLedger_->name;
+    assert(state_ == State::SYNC_FROM_CLOUD && curCtx_ && curCtx_->syncDir == LEDGER_SYNC_DIRECTION_CLOUD_TO_DEVICE &&
+            curCtx_->taskRunning && !stream_);
+    auto name = (const char*)curCtx_->name;
     if (result != 0) {
         LOG(ERROR, "Failed to sync ledger: %s; response result: %d", name, result);
         if (result != PB_CLOUD(Response_Result_LEDGER_NOT_FOUND) &&
@@ -887,14 +674,14 @@ int LedgerManager::receiveGetDataResponse(coap_message* msg, int result, int /* 
         }
         // Ledger is no longer accessible, re-request its info
         LOG(WARN, "Re-requesting ledger info: %s", name);
-        setPendingState(curLedger_, PendingState::GET_INFO | PendingState::SYNC_FROM_CLOUD);
-        curLedger_->taskRunning = false;
-        curLedger_ = nullptr;
+        setPendingState(curCtx_, PendingState::GET_INFO | PendingState::SYNC_FROM_CLOUD);
+        curCtx_->taskRunning = false;
+        curCtx_ = nullptr;
         return 0;
     }
     LOG(TRACE, "Received ledger data: %s", name);
     RefCountPtr<Ledger> ledger;
-    CHECK(getLedger(ledger, name, false /* create */));
+    CHECK(getLedger(ledger, name));
     LedgerWriter writer;
     CHECK(ledger->initWriter(writer, LedgerWriteSource::SYSTEM));
     PB_LEDGER(GetDataResponse) pbResp = {};
@@ -928,14 +715,16 @@ int LedgerManager::receiveGetDataResponse(coap_message* msg, int result, int /* 
         return (d.error < 0) ? d.error : SYSTEM_ERROR_BAD_DATA;
     }
     if (d.hasData) {
-        auto& info = writer.ledgerInfo();
+        LedgerInfo info;
         info.lastUpdated(pbResp.last_updated);
-        auto t = CHECK(getCurrentTime());
-        info.lastSynced(t);
+        auto now = CHECK(getMillisSinceEpoch());
+        info.lastSynced(now);
+        writer.updateInfo(info);
         CHECK(writer.close());
+        ledger->notifySynced(); // TODO: Invoke asynchronously
     }
-    curLedger_->taskRunning = false;
-    curLedger_ = nullptr;
+    curCtx_->taskRunning = false;
+    curCtx_ = nullptr;
     // TODO: Reorder the ledger entries so that they're synchronized in a round-robin fashion
     return 0;
 }
@@ -950,7 +739,7 @@ int LedgerManager::receiveSubscribeResponse(coap_message* msg, int result, int /
         }
         // Some of the cloud-to-device ledgers are no longer accessible, re-request their info
         LOG(WARN, "Re-requesting ledger info");
-        for (auto& ctx: allLedgers_) {
+        for (auto& ctx: contexts_) {
             if (ctx->taskRunning) {
                 setPendingState(ctx.get(), PendingState::GET_INFO);
                 ctx->taskRunning = false;
@@ -974,7 +763,7 @@ int LedgerManager::receiveSubscribeResponse(coap_message* msg, int result, int /
         }
         LOG(TRACE, "Ledger: %s", pbLedger.name);
         bool found = false;
-        auto it = d->self->findLedger(pbLedger.name, found);
+        auto it = d->self->findContext(pbLedger.name, found);
         if (!found) {
             LOG(WARN, "Unknown ledger: %s", pbLedger.name);
             return true; // Ignore
@@ -996,7 +785,7 @@ int LedgerManager::receiveSubscribeResponse(coap_message* msg, int result, int /
     if (!pb_decode(&stream, &PB_LEDGER(SubscribeResponse_msg), &pbResp)) {
         return (d.error < 0) ? d.error : SYSTEM_ERROR_BAD_DATA;
     }
-    for (auto& ctx: allLedgers_) {
+    for (auto& ctx: contexts_) {
         if (ctx->taskRunning) {
             LOG(ERROR, "Missing subscription: %s", (const char*)ctx->name);
             return SYSTEM_ERROR_LEDGER_INVALID_RESPONSE;
@@ -1030,11 +819,11 @@ int LedgerManager::receiveGetInfoResponse(coap_message* msg, int result, int /* 
         LOG(TRACE, "Ledger: %s; scope: %d; sync direction: %d", pbLedger.name, (int)pbLedger.scope,
                 (int)pbLedger.sync_direction);
         RefCountPtr<Ledger> ledger;
-        d->error = d->self->getLedger(ledger, pbLedger.name, false /* create */);
+        d->error = d->self->getLedger(ledger, pbLedger.name);
         if (d->error < 0) {
             return false;
         }
-        auto ctx = static_cast<LedgerContext*>(ledger->context());
+        auto ctx = ledger->syncContext();
         if (!ctx->taskRunning) {
             LOG(ERROR, "Unexpected ledger info: %s", pbLedger.name);
             d->error = SYSTEM_ERROR_LEDGER_INVALID_RESPONSE;
@@ -1058,7 +847,7 @@ int LedgerManager::receiveGetInfoResponse(coap_message* msg, int result, int /* 
                 newInfo.syncDirection(LEDGER_SYNC_DIRECTION_UNKNOWN);
                 newInfo.scope(LEDGER_SCOPE_UNKNOWN);
                 if (ctx->syncDir == LEDGER_SYNC_DIRECTION_CLOUD_TO_DEVICE) { // CLOUD_TO_DEVICE -> DEVICE_TO_CLOUD
-                    ctx->resetDeviceToCloudSyncState();
+                    ctx->resetDeviceToCloudState();
                 }
                 d->localInfoIsInvalid = true; // Will cause a transition to the failed state
             } else if (newInfo.syncDirection() == LEDGER_SYNC_DIRECTION_CLOUD_TO_DEVICE) { // UNKNOWN -> CLOUD_TO_DEVICE
@@ -1067,7 +856,7 @@ int LedgerManager::receiveGetInfoResponse(coap_message* msg, int result, int /* 
                     newInfo.syncPending(false);
                     newInfo.lastUpdated(0);
                 }
-                ctx->resetCloudToDeviceSyncState();
+                ctx->resetCloudToDeviceState();
                 d->resubscribe = true;
             } else if (localInfo.syncPending()) { // UNKNOWN -> DEVICE_TO_CLOUD
                 d->self->setPendingState(ctx, PendingState::SYNC_TO_CLOUD);
@@ -1087,7 +876,7 @@ int LedgerManager::receiveGetInfoResponse(coap_message* msg, int result, int /* 
     if (!pb_decode(&stream, &PB_LEDGER(GetInfoResponse_msg), &pbResp)) {
         return (d.error < 0) ? d.error : SYSTEM_ERROR_BAD_DATA;
     }
-    for (auto& ctx: allLedgers_) {
+    for (auto& ctx: contexts_) {
         if (ctx->taskRunning) {
             // Ledger doesn't exist or is no longer accessible by the device
             LOG(WARN, "Ledger not found: %s", (const char*)ctx->name);
@@ -1096,10 +885,10 @@ int LedgerManager::receiveGetInfoResponse(coap_message* msg, int result, int /* 
                 info.syncDirection(LEDGER_SYNC_DIRECTION_UNKNOWN);
                 info.scope(LEDGER_SCOPE_UNKNOWN);
                 RefCountPtr<Ledger> ledger;
-                CHECK(getLedger(ledger, ctx->name, false /* create */));
+                CHECK(getLedger(ledger, ctx->name));
                 CHECK(ledger->updateInfo(info));
                 if (ctx->syncDir == LEDGER_SYNC_DIRECTION_CLOUD_TO_DEVICE) { // CLOUD_TO_DEVICE -> UNKNOWN
-                    ctx->resetDeviceToCloudSyncState();
+                    ctx->resetDeviceToCloudState();
                     // Don't bother unsubscribing from the deleted ledger
                 }
                 ctx->syncDir = LEDGER_SYNC_DIRECTION_UNKNOWN;
@@ -1116,25 +905,20 @@ int LedgerManager::receiveGetInfoResponse(coap_message* msg, int result, int /* 
     return 0;
 }
 
-int LedgerManager::sendSetDataRequest(LedgerContext* ctx) {
+int LedgerManager::sendSetDataRequest(LedgerSyncContext* ctx) {
     assert(state_ == State::READY && (ctx->pendingState & PendingState::SYNC_TO_CLOUD) &&
-            ctx->syncDir == LEDGER_SYNC_DIRECTION_DEVICE_TO_CLOUD && !curLedger_ && !stream_);
+            ctx->syncDir == LEDGER_SYNC_DIRECTION_DEVICE_TO_CLOUD && !curCtx_ && !stream_);
     auto name = (const char*)ctx->name;
     RefCountPtr<Ledger> ledger;
-    CHECK(getLedger(ledger, name, false /* create */));
-    // Create a reader and get the current ledger info atomically
-    std::unique_lock lock(*ledger);
+    CHECK(getLedger(ledger, name));
     LedgerReader reader;
     CHECK(ledger->initReader(reader));
-    auto info = ledger->info();
-    lock.unlock();
     struct EncodeContext {
         LedgerManager* self;
         LedgerReader* reader;
-        LedgerInfo* info;
         int error;
     };
-    EncodeContext d = { .self = this, .reader = &reader, .info = &info, .error = 0 };
+    EncodeContext d = { .self = this, .reader = &reader, .error = 0 };
     PB_CLOUD(Request) pbReq = {};
     pbReq.type = PB_CLOUD(Request_Type_LEDGER_SET_DATA);
     pbReq.which_data = PB_CLOUD(Request_ledger_set_data_tag);
@@ -1142,11 +926,11 @@ int LedgerManager::sendSetDataRequest(LedgerContext* ctx) {
     if (n >= sizeof(pbReq.data.ledger_set_data.name)) {
         return SYSTEM_ERROR_INTERNAL;
     }
-    pbReq.data.ledger_set_data.last_updated = info.lastUpdated();
+    pbReq.data.ledger_set_data.last_updated = reader.info().lastUpdated();
     pbReq.data.ledger_set_data.data.arg = &d;
     pbReq.data.ledger_set_data.data.funcs.encode = [](pb_ostream_t* stream, const pb_field_iter_t* field, void* const* arg) {
         auto d = (EncodeContext*)*arg;
-        size_t size = d->info->dataSize();
+        size_t size = d->reader->info().dataSize();
         if (!pb_encode_tag_for_field(stream, field) || !pb_encode_varint(stream, size)) {
             return false;
         }
@@ -1184,15 +968,15 @@ int LedgerManager::sendSetDataRequest(LedgerContext* ctx) {
     // Clear the pending state
     clearPendingState(ctx, PendingState::SYNC_TO_CLOUD);
     ctx->taskRunning = true;
-    curLedger_ = ctx;
+    curCtx_ = ctx;
     reqId_ = reqId;
     state_ = State::SYNC_TO_CLOUD;
     return 0;
 }
 
-int LedgerManager::sendGetDataRequest(LedgerContext* ctx) {
+int LedgerManager::sendGetDataRequest(LedgerSyncContext* ctx) {
     assert(state_ == State::READY && (ctx->pendingState & PendingState::SYNC_FROM_CLOUD) &&
-            ctx->syncDir == LEDGER_SYNC_DIRECTION_CLOUD_TO_DEVICE && !curLedger_);
+            ctx->syncDir == LEDGER_SYNC_DIRECTION_CLOUD_TO_DEVICE && !curCtx_);
     PB_CLOUD(Request) pbReq = {};
     pbReq.type = PB_CLOUD(Request_Type_LEDGER_GET_DATA);
     pbReq.which_data = PB_CLOUD(Request_ledger_get_data_tag);
@@ -1220,7 +1004,7 @@ int LedgerManager::sendGetDataRequest(LedgerContext* ctx) {
     // Clear the pending state
     clearPendingState(ctx, PendingState::SYNC_FROM_CLOUD);
     ctx->taskRunning = true;
-    curLedger_ = ctx;
+    curCtx_ = ctx;
     reqId_ = reqId;
     state_ = State::SYNC_FROM_CLOUD;
     return 0;
@@ -1235,7 +1019,7 @@ int LedgerManager::sendSubscribeRequest() {
     pbReq.data.ledger_subscribe.ledgers.funcs.encode = [](pb_ostream_t* stream, const pb_field_iter_t* field, void* const* arg) {
         // Make sure not to update any state in this callback as it may be called multiple times
         auto self = (const LedgerManager*)*arg;
-        for (auto& ctx: self->allLedgers_) {
+        for (auto& ctx: self->contexts_) {
             if (ctx->pendingState & PendingState::SUBSCRIBE) {
                 if (!pb_encode_tag_for_field(stream, field) ||
                         !pb_encode_string(stream, (const pb_byte_t*)(const char*)ctx->name, std::strlen(ctx->name))) {
@@ -1259,7 +1043,7 @@ int LedgerManager::sendSubscribeRequest() {
     destroyMsgGuard.dismiss();
     // Clear the pending state
     pendingState_ = 0;
-    for (auto& ctx: allLedgers_) {
+    for (auto& ctx: contexts_) {
         if (ctx->pendingState & PendingState::SUBSCRIBE) {
             ctx->pendingState &= ~PendingState::SUBSCRIBE;
             ctx->taskRunning = true;
@@ -1280,7 +1064,7 @@ int LedgerManager::sendGetInfoRequest() {
     pbReq.data.ledger_get_info.ledgers.funcs.encode = [](pb_ostream_t* stream, const pb_field_iter_t* field, void* const* arg) {
         // Make sure not to update any state in this callback as it may be called multiple times
         auto self = (const LedgerManager*)*arg;
-        for (auto& ctx: self->allLedgers_) {
+        for (auto& ctx: self->contexts_) {
             if (ctx->pendingState & PendingState::GET_INFO) {
                 if (!pb_encode_tag_for_field(stream, field) ||
                         !pb_encode_string(stream, (const pb_byte_t*)(const char*)ctx->name, std::strlen(ctx->name))) {
@@ -1304,7 +1088,7 @@ int LedgerManager::sendGetInfoRequest() {
     destroyMsgGuard.dismiss();
     // Clear the pending state
     pendingState_ = 0;
-    for (auto& ctx: allLedgers_) {
+    for (auto& ctx: contexts_) {
         if (ctx->pendingState & PendingState::GET_INFO) {
             ctx->pendingState &= ~PendingState::GET_INFO;
             ctx->taskRunning = true;
@@ -1335,9 +1119,8 @@ int LedgerManager::sendResponse(int result, int reqId) {
     return 0;
 }
 
-void LedgerManager::ledgerUpdated(void* context) {
+void LedgerManager::notifyLedgerChanged(LedgerSyncContext* ctx) {
     std::lock_guard lock(mutex_);
-    auto ctx = static_cast<LedgerContext*>(context);
     if (ctx->syncDir == LEDGER_SYNC_DIRECTION_DEVICE_TO_CLOUD || ctx->syncDir == LEDGER_SYNC_DIRECTION_UNKNOWN) {
         // Mark the ledger as changed but only schedule a sync for it if its actual sync direction is known
         ctx->changed = true;
@@ -1348,20 +1131,20 @@ void LedgerManager::ledgerUpdated(void* context) {
     }
 }
 
-void LedgerManager::setPendingState(LedgerContext* ctx, int state) {
+void LedgerManager::setPendingState(LedgerSyncContext* ctx, int state) {
     ctx->pendingState |= state;
     pendingState_ |= state;
 }
 
-void LedgerManager::clearPendingState(LedgerContext* ctx, int state) {
+void LedgerManager::clearPendingState(LedgerSyncContext* ctx, int state) {
     ctx->pendingState &= ~state;
     pendingState_ = 0;
-    for (auto& ctx: allLedgers_) {
+    for (auto& ctx: contexts_) {
         pendingState_ |= ctx->pendingState;
     }
 }
 
-void LedgerManager::updateSyncTime(LedgerContext* ctx) {
+void LedgerManager::updateSyncTime(LedgerSyncContext* ctx) {
     auto now = hal_timer_millis(nullptr);
     if (!ctx->forcedSyncTime) {
         ctx->forcedSyncTime = now + MAX_SYNC_DELAY;
@@ -1379,9 +1162,9 @@ void LedgerManager::handleError(int error) {
     }
 }
 
-LedgerManager::LedgerContexts::ConstIterator LedgerManager::findLedger(const char* name, bool& found) const {
+LedgerManager::LedgerSyncContexts::ConstIterator LedgerManager::findContext(const char* name, bool& found) const {
     found = false;
-    auto it = std::lower_bound(allLedgers_.begin(), allLedgers_.end(), name, [&found](const auto& ctx, const char* name) {
+    auto it = std::lower_bound(contexts_.begin(), contexts_.end(), name, [&found](const auto& ctx, const char* name) {
         auto r = std::strcmp(ctx->name, name);
         if (r == 0) {
             found = true;
@@ -1399,7 +1182,7 @@ void LedgerManager::addLedgerRef(const LedgerBase* ledger) {
 void LedgerManager::releaseLedger(const LedgerBase* ledger) {
     std::lock_guard lock(mutex_);
     if (--ledger->refCount() == 0) {
-        auto ctx = static_cast<LedgerContext*>(ledger->context());
+        auto ctx = ledger->syncContext();
         ctx->instance = nullptr;
         delete ledger;
     }
@@ -1477,593 +1260,6 @@ void LedgerManager::requestErrorCallback(int error, int reqId, void* arg) {
 LedgerManager* LedgerManager::instance() {
     static LedgerManager mgr;
     return &mgr;
-}
-
-Ledger::Ledger(void* ctx) :
-        LedgerBase(ctx),
-        lastSeqNum_(0),
-        stagedSeqNum_(0),
-        curReaderCount_(0),
-        stagedReaderCount_(0),
-        stagedFileCount_(0),
-        lastUpdated_(0),
-        lastSynced_(0),
-        dataSize_(0),
-        syncPending_(false),
-        syncCallback_(nullptr),
-        destroyAppData_(nullptr),
-        appData_(nullptr),
-        name_(""),
-        scope_(LEDGER_SCOPE_UNKNOWN),
-        syncDir_(LEDGER_SYNC_DIRECTION_UNKNOWN),
-        inited_(false) {
-}
-
-Ledger::~Ledger() {
-    if (appData_ && destroyAppData_) {
-        destroyAppData_(appData_);
-    }
-}
-
-int Ledger::init(const char* name) {
-    if (!isLedgerNameValid(name)) {
-        return SYSTEM_ERROR_INVALID_ARGUMENT;
-    }
-    name_ = name;
-    FsLock fs;
-    int r = CHECK(loadLedgerInfo(fs.instance()));
-    if (r == RESULT_CURRENT_DATA_NOT_FOUND) {
-        // Initialize the ledger directory
-        char path[MAX_PATH_LEN + 1];
-        CHECK(getTempDirPath(path, sizeof(path), name_));
-        CHECK(mkdirp(path));
-        CHECK(getStagedDirPath(path, sizeof(path), name_));
-        CHECK(mkdirp(path));
-        // Create a file for the current ledger data
-        CHECK(initCurrentData(fs.instance()));
-    } else {
-        // Clean up and recover staged data
-        CHECK(removeTempData(fs.instance()));
-        r = CHECK(flushStagedData(fs.instance()));
-        if (r > 0) {
-            // Found staged data. Reload the ledger info
-            CHECK(loadLedgerInfo(fs.instance()));
-        }
-    }
-    inited_ = true;
-    return 0;
-}
-
-int Ledger::initReader(LedgerReader& reader) {
-    std::lock_guard lock(*this);
-    if (!inited_) {
-        return SYSTEM_ERROR_INVALID_STATE;
-    }
-    CHECK(reader.init(dataSize_, stagedSeqNum_, this));
-    if (stagedSeqNum_ > 0) {
-        ++stagedReaderCount_;
-    } else {
-        ++curReaderCount_;
-    }
-    return 0;
-}
-
-int Ledger::initWriter(LedgerWriter& writer, LedgerWriteSource src) {
-    std::lock_guard lock(*this);
-    if (!inited_) {
-        return SYSTEM_ERROR_INVALID_STATE;
-    }
-    // It's allowed to write to a ledger while its sync direction is unknown
-    if (src == LedgerWriteSource::USER && syncDir_ != LEDGER_SYNC_DIRECTION_DEVICE_TO_CLOUD &&
-            syncDir_ != LEDGER_SYNC_DIRECTION_UNKNOWN) {
-        return SYSTEM_ERROR_LEDGER_READ_ONLY;
-    }
-    CHECK(writer.init(src, ++lastSeqNum_, this));
-    return 0;
-}
-
-LedgerInfo Ledger::info() const {
-    std::lock_guard lock(*this);
-    return LedgerInfo()
-            .scope(scope_)
-            .syncDirection(syncDir_)
-            .dataSize(dataSize_)
-            .lastUpdated(lastUpdated_)
-            .lastSynced(lastSynced_)
-            .syncPending(syncPending_);
-}
-
-int Ledger::updateInfo(const LedgerInfo& info) {
-    std::lock_guard lock(*this);
-    if (!inited_) {
-        return SYSTEM_ERROR_INVALID_STATE;
-    }
-    // Open the file with the current ledger data
-    char path[MAX_PATH_LEN + 1];
-    CHECK(getCurrentFilePath(path, sizeof(path), name_));
-    FsLock fs;
-    lfs_file_t file = {};
-    // TODO: Rewriting parts of a file is inefficient with LittleFS. Consider using an append-only
-    // structure for storing ledger data
-    CHECK_FS(lfs_file_open(fs.instance(), &file, path, LFS_O_RDWR));
-    NAMED_SCOPE_GUARD(closeFileGuard, {
-        int r = closeFile(fs.instance(), &file);
-        if (r < 0) {
-            LOG(ERROR, "Error while closing file: %d", r);
-        }
-    });
-    // Read the footer
-    size_t dataSize = 0;
-    size_t infoSize = 0;
-    CHECK_FS(lfs_file_seek(fs.instance(), &file, -(int)sizeof(LedgerDataFooter), LFS_SEEK_END));
-    CHECK(readFooter(fs.instance(), &file, &dataSize, &infoSize));
-    // Write the info section
-    CHECK_FS(lfs_file_seek(fs.instance(), &file, -(int)(infoSize + sizeof(LedgerDataFooter)), LFS_SEEK_END));
-    auto newInfo = this->info().update(info);
-    size_t newInfoSize = CHECK(writeLedgerInfo(fs.instance(), &file, name_, newInfo));
-    // Write the footer
-    if (newInfoSize != infoSize) {
-        CHECK(writeFooter(fs.instance(), &file, dataSize, newInfoSize));
-        size_t newFileSize = CHECK_FS(lfs_file_tell(fs.instance(), &file));
-        CHECK_FS(lfs_file_truncate(fs.instance(), &file, newFileSize));
-    }
-    closeFileGuard.dismiss();
-    CHECK_FS(lfs_file_close(fs.instance(), &file));
-    // Update the instance
-    setLedgerInfo(newInfo);
-    return 0;
-}
-
-int Ledger::readerClosed(bool staged) {
-    std::lock_guard lock(*this);
-    if (staged) {
-        --stagedReaderCount_;
-    } else {
-        --curReaderCount_;
-    }
-    // Check if there's staged data and if it can be flushed
-    int result = 0;
-    if (stagedSeqNum_ > 0 && curReaderCount_ == 0 && stagedReaderCount_ == 0) {
-        // To save on RAM usage, we don't count how many readers are open for each individual file,
-        // which may result in the staged directory containing multiple unused files with outdated
-        // ledger data. Typically, however, when the last reader is closed, there will be at most
-        // one file with staged data as the application never keeps ledger streams open for long.
-        // The only slow reader is the system which streams ledger data to the server
-        FsLock fs;
-        if (stagedFileCount_ == 1) {
-            // Fast track: there's one file with staged data. Move it to "current"
-            char srcPath[MAX_PATH_LEN + 1];
-            char destPath[MAX_PATH_LEN + 1];
-            CHECK(getStagedFilePath(srcPath, sizeof(srcPath), name_, stagedSeqNum_));
-            CHECK(getCurrentFilePath(destPath, sizeof(destPath), name_));
-            result = renameFile(fs.instance(), srcPath, destPath);
-            if (result < 0) {
-                // The filesystem is full or broken, or somebody messed up the staged directory
-                LOG(ERROR, "Failed to rename file: %d", result);
-                int r = removeFile(fs.instance(), srcPath);
-                if (r < 0) {
-                    LOG(WARN, "Failed to remove file: %d", r);
-                }
-            }
-        } else {
-            // The staged directory contains outdated files. Clean those up and move the most recent
-            // file to "current"
-            int r = flushStagedData(fs.instance());
-            if (r < 0) {
-                result = r;
-            } else if (r != stagedSeqNum_) {
-                // The file moved wasn't the most recent one for some reason
-                LOG(ERROR, "Staged ledger data not found");
-                result = SYSTEM_ERROR_LEDGER_INCONSISTENT_STATE;
-            }
-        }
-        if (result < 0) {
-            // Try recovering some known consistent state
-            int r = loadLedgerInfo(fs.instance());
-            if (r < 0) {
-                LOG(ERROR, "Failed to recover ledger state");
-            }
-        }
-        stagedSeqNum_ = 0;
-        stagedFileCount_ = 0;
-    }
-    return result;
-}
-
-int Ledger::writerClosed(const LedgerInfo& info, LedgerWriteSource src, int tempSeqNum) {
-    std::unique_lock lock(*this);
-    FsLock fs;
-    // Move the file where appropriate
-    bool newStagedFile = false;
-    char srcPath[MAX_PATH_LEN + 1];
-    char destPath[MAX_PATH_LEN + 1];
-    CHECK(getTempFilePath(srcPath, sizeof(srcPath), name_, tempSeqNum));
-    if (curReaderCount_ == 0 && stagedReaderCount_ == 0) {
-        // Nobody is reading anything. Move to "current"
-        CHECK(getCurrentFilePath(destPath, sizeof(destPath), name_));
-    } else if (stagedSeqNum_ > 0 && stagedReaderCount_ == 0) {
-        // There's staged data and nobody is reading it. Replace it
-        CHECK(getStagedFilePath(destPath, sizeof(destPath), name_, stagedSeqNum_));
-    } else {
-        // Create a new file in "staged". Note that in this case, "current" can't be replaced with
-        // the new data even if nobody is reading it, as otherwise it would get overwritten with
-        // outdated data if the device resets before the staged directory is cleaned up
-        CHECK(getStagedFilePath(destPath, sizeof(destPath), name_, tempSeqNum));
-        newStagedFile = true;
-    }
-    CHECK_FS(lfs_rename(fs.instance(), srcPath, destPath));
-    if (newStagedFile) {
-        stagedSeqNum_ = tempSeqNum;
-        ++stagedFileCount_;
-    }
-    setLedgerInfo(this->info().update(info));
-    if (src == LedgerWriteSource::USER) {
-        // Unlock the ledger before calling into the manager to avoid a deadlock
-        lock.unlock();
-        fs.unlock();
-        LedgerManager::instance()->ledgerUpdated(context());
-        fs.lock(); // FIXME
-    } else if (syncCallback_) { // SYSTEM
-        syncCallback_(reinterpret_cast<ledger_instance*>(this), appData_);
-    }
-    return 0;
-}
-
-int Ledger::loadLedgerInfo(lfs_t* fs) {
-    // Open the file with the current ledger data
-    char path[MAX_PATH_LEN + 1];
-    CHECK(getCurrentFilePath(path, sizeof(path), name_));
-    lfs_file_t file = {};
-    int r = lfs_file_open(fs, &file, path, LFS_O_RDONLY);
-    if (r < 0) {
-        if (r == LFS_ERR_NOENT) {
-            return RESULT_CURRENT_DATA_NOT_FOUND;
-        }
-        CHECK_FS(r); // Forward the error
-    }
-    SCOPE_GUARD({
-        int r = closeFile(fs, &file);
-        if (r < 0) {
-            LOG(ERROR, "Error while closing file: %d", r);
-        }
-    });
-    // Read the footer
-    size_t dataSize = 0;
-    size_t infoSize = 0;
-    int version = 0;
-    CHECK_FS(lfs_file_seek(fs, &file, -(int)sizeof(LedgerDataFooter), LFS_SEEK_END));
-    CHECK(readFooter(fs, &file, &dataSize, &infoSize, &version));
-    if (version != DATA_FORMAT_VERSION) {
-        LOG(ERROR, "Unsupported version of ledger data format: %d", version);
-        return SYSTEM_ERROR_LEDGER_UNSUPPORTED_FORMAT;
-    }
-    size_t fileSize = CHECK_FS(lfs_file_size(fs, &file));
-    if (fileSize != dataSize + infoSize + sizeof(LedgerDataFooter)) {
-        LOG(ERROR, "Unexpected size of ledger data file");
-        return SYSTEM_ERROR_LEDGER_INVALID_FORMAT;
-    }
-    // Skip the data section
-    CHECK_FS(lfs_file_seek(fs, &file, dataSize, LFS_SEEK_SET));
-    // Read the info section
-    PB_INTERNAL(LedgerInfo) pbInfo = {};
-    r = decodeProtobufFromFile(&file, &PB_INTERNAL(LedgerInfo_msg), &pbInfo, infoSize);
-    if (r < 0) {
-        LOG(ERROR, "Failed to parse ledger info: %d", r);
-        return SYSTEM_ERROR_LEDGER_INVALID_FORMAT;
-    }
-    if (std::strcmp(pbInfo.name, name_) != 0) {
-        LOG(ERROR, "Unexpected ledger name");
-        return SYSTEM_ERROR_LEDGER_INVALID_FORMAT;
-    }
-    scope_ = static_cast<ledger_scope>(pbInfo.scope);
-    syncDir_ = static_cast<ledger_sync_direction>(pbInfo.sync_direction);
-    dataSize_ = dataSize;
-    lastUpdated_ = pbInfo.last_updated;
-    lastSynced_ = pbInfo.last_synced;
-    syncPending_ = pbInfo.sync_pending;
-    return 0;
-}
-
-void Ledger::setLedgerInfo(const LedgerInfo& info) {
-    assert(info.hasScope() && info.hasSyncDirection() && info.hasDataSize() && info.hasLastUpdated() &&
-            info.hasLastSynced() && info.hasSyncPending());
-    scope_ = info.scope();
-    syncDir_ = info.syncDirection();
-    dataSize_ = info.dataSize();
-    lastUpdated_ = info.lastUpdated();
-    lastSynced_ = info.lastSynced();
-    syncPending_ = info.syncPending();
-}
-
-int Ledger::initCurrentData(lfs_t* fs) {
-    // Create a file for the current ledger data
-    char path[MAX_PATH_LEN + 1];
-    CHECK(getCurrentFilePath(path, sizeof(path), name_));
-    lfs_file_t file = {};
-    CHECK_FS(lfs_file_open(fs, &file, path, LFS_O_WRONLY | LFS_O_CREAT | LFS_O_EXCL));
-    SCOPE_GUARD({
-        int r = closeFile(fs, &file);
-        if (r < 0) {
-            LOG(ERROR, "Error while closing file: %d", r);
-        }
-    });
-    // Write the info section
-    size_t infoSize = CHECK(writeLedgerInfo(fs, &file, name_, info()));
-    // Write the footer
-    CHECK(writeFooter(fs, &file, 0 /* dataSize */, infoSize));
-    return 0;
-}
-
-int Ledger::flushStagedData(lfs_t* fs) {
-    char path[MAX_PATH_LEN + 1];
-    CHECK(getStagedDirPath(path, sizeof(path), name_));
-    lfs_dir_t dir = {};
-    CHECK_FS(lfs_dir_open(fs, &dir, path));
-    SCOPE_GUARD({
-        int r = closeDir(fs, &dir);
-        if (r < 0) {
-            LOG(ERROR, "Failed to close directory handle: %d", r);
-        }
-    });
-    // Find the latest staged data
-    char fileName[32] = {};
-    int maxSeqNum = 0;
-    lfs_info entry = {};
-    int r = 0;
-    while ((r = lfs_dir_read(fs, &dir, &entry)) == 1) {
-        if (entry.type != LFS_TYPE_REG) {
-            if (entry.type != LFS_TYPE_DIR || (std::strcmp(entry.name, ".") != 0 && std::strcmp(entry.name, "..") != 0)) {
-                LOG(WARN, "Found unexpected entry in ledger directory");
-            }
-            continue;
-        }
-        char* end = nullptr;
-        int seqNum = std::strtol(entry.name, &end, 10);
-        if (end != entry.name + std::strlen(entry.name)) {
-            LOG(WARN, "Found unexpected entry in ledger directory");
-            continue;
-        }
-        if (seqNum > maxSeqNum) {
-            if (maxSeqNum > 0) {
-                // Older staged files must be removed before the most recent file is moved to
-                // "current", otherwise "current" would get overwritten with outdated data if
-                // the device resets before the staged directory is cleaned up
-                CHECK(getStagedFilePath(path, sizeof(path), name_, fileName));
-                CHECK_FS(lfs_remove(fs, path));
-            }
-            size_t n = strlcpy(fileName, entry.name, sizeof(fileName));
-            if (n >= sizeof(fileName)) {
-                return SYSTEM_ERROR_INTERNAL;
-            }
-            maxSeqNum = seqNum;
-        } else {
-            CHECK(getStagedFilePath(path, sizeof(path), name_, entry.name));
-            CHECK_FS(lfs_remove(fs, path));
-        }
-    }
-    CHECK_FS(r);
-    if (maxSeqNum > 0) {
-        // Replace "current" with the found file
-        char destPath[MAX_PATH_LEN + 1];
-        CHECK(getCurrentFilePath(destPath, sizeof(destPath), name_));
-        CHECK(getStagedFilePath(path, sizeof(path), name_, fileName));
-        CHECK_FS(lfs_rename(fs, path, destPath));
-        return maxSeqNum;
-    }
-    return 0;
-}
-
-int Ledger::removeTempData(lfs_t* fs) {
-    char path[MAX_PATH_LEN + 1];
-    CHECK(getTempDirPath(path, sizeof(path), name_));
-    lfs_dir_t dir = {};
-    CHECK_FS(lfs_dir_open(fs, &dir, path));
-    SCOPE_GUARD({
-        int r = closeDir(fs, &dir);
-        if (r < 0) {
-            LOG(ERROR, "Failed to close directory handle: %d", r);
-        }
-    });
-    lfs_info entry = {};
-    int r = 0;
-    while ((r = lfs_dir_read(fs, &dir, &entry)) == 1) {
-        if (entry.type != LFS_TYPE_REG) {
-            if (entry.type != LFS_TYPE_DIR || (std::strcmp(entry.name, ".") != 0 && std::strcmp(entry.name, "..") != 0)) {
-                LOG(WARN, "Found unexpected entry in ledger directory");
-            }
-            continue;
-        }
-        CHECK(getTempFilePath(path, sizeof(path), name_, entry.name));
-        CHECK_FS(lfs_remove(fs, path));
-    }
-    CHECK_FS(r);
-    return 0;
-}
-
-LedgerInfo& LedgerInfo::update(const LedgerInfo& info) {
-    if (info.hasScope()) {
-        scope_ = info.scope();
-    }
-    if (info.hasSyncDirection()) {
-        syncDir_ = info.syncDirection();
-    }
-    if (info.hasDataSize()) {
-        dataSize_ = info.dataSize();
-    }
-    if (info.hasLastUpdated()) {
-        lastUpdated_ = info.lastUpdated();
-    }
-    if (info.hasLastSynced()) {
-        lastSynced_ = info.lastSynced();
-    }
-    if (info.hasSyncPending()) {
-        syncPending_ = info.syncPending();
-    }
-    return *this;
-}
-
-int LedgerReader::init(size_t dataSize, int stagedSeqNum, Ledger* ledger) {
-    char path[MAX_PATH_LEN + 1];
-    if (stagedSeqNum > 0) {
-        // The most recent data is staged
-        CHECK(getStagedFilePath(path, sizeof(path), ledger->name(), stagedSeqNum));
-        staged_ = true;
-    } else {
-        CHECK(getCurrentFilePath(path, sizeof(path), ledger->name()));
-    }
-    // Open the respective file
-    FsLock fs;
-    CHECK_FS(lfs_file_open(fs.instance(), &file_, path, LFS_O_RDONLY));
-    ledger_ = ledger;
-    dataSize_ = dataSize;
-    open_ = true;
-    return 0;
-}
-
-int LedgerReader::read(char* data, size_t size) {
-    if (!open_) {
-        return SYSTEM_ERROR_INVALID_STATE;
-    }
-    size_t bytesToRead = std::min(size, dataSize_ - dataOffs_);
-    if (size > 0 && bytesToRead == 0) {
-        return SYSTEM_ERROR_END_OF_STREAM;
-    }
-    FsLock fs;
-    size_t n = CHECK_FS(lfs_file_read(fs.instance(), &file_, data, bytesToRead));
-    if (n != bytesToRead) {
-        LOG(ERROR, "Unexpected end of ledger data file");
-        return SYSTEM_ERROR_LEDGER_INVALID_FORMAT;
-    }
-    dataOffs_ += n;
-    return n;
-}
-
-int LedgerReader::close(bool /* discard */) {
-    if (!open_) {
-        return 0;
-    }
-    // Consider the reader closed regardless if any of the operations below fails
-    open_ = false;
-    FsLock fs;
-    int result = closeFile(fs.instance(), &file_);
-    if (result < 0) {
-        LOG(ERROR, "Error while closing file: %d", result);
-    }
-    // Let the ledger flush any data
-    int r = ledger_->readerClosed(staged_);
-    if (r < 0) {
-        LOG(ERROR, "Failed to flush ledger data: %d", r);
-        return r;
-    }
-    return result;
-}
-
-int LedgerReader::rewind() {
-    if (!open_) {
-        return SYSTEM_ERROR_INVALID_STATE;
-    }
-    FsLock fs;
-    CHECK_FS(lfs_file_seek(fs.instance(), &file_, 0, LFS_SEEK_SET));
-    dataOffs_ = 0;
-    return 0;
-}
-
-int LedgerWriter::init(LedgerWriteSource src, int tempSeqNum, Ledger* ledger) {
-    // Create a temporary file
-    char path[MAX_PATH_LEN + 1];
-    CHECK(getTempFilePath(path, sizeof(path), ledger->name(), tempSeqNum));
-    FsLock fs;
-    CHECK_FS(lfs_file_open(fs.instance(), &file_, path, LFS_O_WRONLY | LFS_O_CREAT | LFS_O_EXCL));
-    ledger_ = ledger;
-    tempSeqNum_ = tempSeqNum;
-    src_ = src;
-    open_ = true;
-    return 0;
-}
-
-int LedgerWriter::write(const char* data, size_t size) {
-    if (!open_) {
-        return SYSTEM_ERROR_INVALID_STATE;
-    }
-    if (src_ == LedgerWriteSource::USER && dataSize_ + size > LEDGER_MAX_DATA_SIZE) {
-        LOG(ERROR, "Ledger data is too long");
-        return SYSTEM_ERROR_LEDGER_TOO_LARGE;
-    }
-    FsLock fs;
-    size_t n = CHECK_FS(lfs_file_write(fs.instance(), &file_, data, size));
-    if (n != size) {
-        LOG(ERROR, "Unexpected number of bytes written");
-        return SYSTEM_ERROR_FILESYSTEM;
-    }
-    dataSize_ += n;
-    return n;
-}
-
-int LedgerWriter::close(bool discard) {
-    if (!open_) {
-        return 0;
-    }
-    // Consider the writer closed regardless if any of the operations below fails
-    open_ = false;
-    FsLock fs;
-    if (discard) {
-        // Remove the temporary file
-        char path[MAX_PATH_LEN + 1];
-        CHECK(getTempFilePath(path, sizeof(path), ledger_->name(), tempSeqNum_));
-        int r = closeFile(fs.instance(), &file_);
-        if (r < 0) {
-            LOG(ERROR, "Error while closing file: %d", r);
-        }
-        CHECK_FS(lfs_remove(fs.instance(), path));
-        return r;
-    }
-    NAMED_SCOPE_GUARD(removeFileGuard, {
-        char path[MAX_PATH_LEN + 1];
-        int r = getTempFilePath(path, sizeof(path), ledger_->name(), tempSeqNum_);
-        if (r >= 0) {
-            r = removeFile(fs.instance(), path);
-        }
-        if (r < 0) {
-            LOG(ERROR, "Failed to remove file: %d", r);
-        }
-    });
-    NAMED_SCOPE_GUARD(closeFileGuard, { // Will run before removeFileGuard
-        int r = closeFile(fs.instance(), &file_);
-        if (r < 0) {
-            LOG(ERROR, "Error while closing file: %d", r);
-        }
-    });
-    // Prepare the updated ledger info
-    std::lock_guard lock(*ledger_);
-    auto newInfo = ledger_->info().update(ledgerInfo_);
-    newInfo.dataSize(dataSize_); // Can't be overridden
-    if (!ledgerInfo_.hasLastUpdated()) {
-        int64_t t = getCurrentTime();
-        if (t < 0) {
-            if (t != SYSTEM_ERROR_HAL_RTC_INVALID_TIME) {
-                LOG(ERROR, "Failed to get current time: %d", (int)t);
-            }
-            t = 0; // Time is unknown
-        }
-        newInfo.lastUpdated(t);
-    }
-    if (!ledgerInfo_.hasSyncPending() && src_ == LedgerWriteSource::USER) {
-        newInfo.syncPending(true);
-    }
-    // Write the info section
-    size_t infoSize = CHECK(writeLedgerInfo(fs.instance(), &file_, ledger_->name(), newInfo));
-    // Write the footer
-    CHECK(writeFooter(fs.instance(), &file_, dataSize_, infoSize));
-    closeFileGuard.dismiss();
-    CHECK_FS(lfs_file_close(fs.instance(), &file_));
-    // Flush the data
-    int r = ledger_->writerClosed(newInfo, src_, tempSeqNum_);
-    if (r < 0) {
-        LOG(ERROR, "Failed to flush ledger data: %d", r);
-        return r;
-    }
-    removeFileGuard.dismiss();
-    return 0;
 }
 
 } // namespace system
