+++ system/src/system_ledger.cpp
@@ -22,10 +22,11 @@
 #include <memory>
 #include <mutex>
 
+#include "ledger/ledger.h"
+#include "ledger/ledger_manager.h"
 #include "system_ledger.h"
-
-#include "system_ledger_internal.h"
 #include "system_threading.h"
+
 #include "check.h"
 
 using namespace particle;
@@ -33,7 +34,7 @@ using namespace particle::system;
 
 int ledger_get_instance(ledger_instance** ledger, const char* name, void* reserved) {
     RefCountPtr<Ledger> lr;
-    CHECK(LedgerManager::instance()->getLedger(lr, name));
+    CHECK(LedgerManager::instance()->getLedger(lr, name, true /* create */));
     *ledger = reinterpret_cast<ledger_instance*>(lr.unwrap()); // Transfer ownership to the caller
     return 0;
 }
